Вынес выбор меню в отдельный класс.
