+++ 6_Lesson/Program.cs
@@ -4,17 +4,19 @@ bool flag = true;
 
 while (flag)
 {
-    //Меню выбора решения
+   
+    
     Console.Clear();
     Console.WriteLine("Выберите решение задачи:");
     Console.WriteLine("1- Решение задачи ДЗ № 1:");
     Console.WriteLine("2- Решение задачи ДЗ № 2:");
     Console.WriteLine("0- выход из программы:");
 
-    int numMenu = Numbers("Введите номер: ");
+    
+    int numMenu = NumbersMenu.Numbers("Введите номер: ");
     //Конец меню
 
-    //Выбор решения задания
+    
     switch (numMenu)
     {
         case 0:
@@ -49,8 +51,7 @@ while (flag)
             }
     }
 
-    //Конец выбора заадния
-    //Условия для выхода из программы
+    
     if (numMenu == 0)
     {
         return;
@@ -60,18 +61,3 @@ while (flag)
 
 Console.WriteLine("Для выхода из программы нажмите любую кнопку...");
 Console.ReadLine();
-
-static int Numbers(string message)
-{
-    do
-    {
-        Console.Write(message);
-        string text = Console.ReadLine();
-        int result;
-        if (int.TryParse(text, out result))
-        {
-            return result;
-        }
-    }
-    while (true);
-}
